requirements.txt
aiortc==1.9.0
aioice==0.9.0
av==12.3.0
numpy>=1.26.0
pillow>=10.3.0      
websockets>=12.0
